fixed tensorflow errors for fedavg

+++ workers/trainer.py
@@ -1,6 +1,6 @@
 import tensorflow as tf
 import numpy as np
-from tensorflow.python.keras.backend import gradients
+from tensorflow.python.keras.backend import dtype, gradients
 from src import config, noise, replaybuffer, environment, util
 from agent import model, ddpgagent
 from workers import evaluator
@@ -367,7 +367,7 @@ class Trainer:
         Returns:
             float: the weight to use for performing weighted averaging
         """
-        return 1/np.mean(self.all_ep_reward_lists[idx1][idx2][-self.conf.weighted_window:]) # calculate the metric for weighting
+        return 1/np.mean(self.all_ep_reward_lists[idx1][idx2][-self.conf.weighted_window:], dtype=np.float32) # calculate the metric for weighting
 
     def compute_weighted_params(self, params, weight):
         return np.multiply(params, weight)
